requirements/base.txt
Django==1.5.2
South==0.8.2
argparse==1.2.1
distribute==0.6.27
wsgiref==0.1.2
PIL==1.1.7
django-easy-maps==0.8.4
geopy==0.95.1
django-registration==1.0
django-registration-bootstrap==0.2.0
django-crispy-forms==1.3.2